+++ django_mysql/models/__init__.py
@@ -8,7 +8,7 @@ from django_mysql.models.query import (  # noqa
     pt_visual_explain, QuerySet, QuerySetMixin
 )
 from django_mysql.models.fields import (  # noqa
-    Bit1BooleanField, DynamicField, EnumField, ListCharField, ListTextField,
-    NullBit1BooleanField, SetCharField, SetTextField, SizedBinaryField,
-    SizedTextField,
+    Bit1BooleanField, DynamicField, EnumField, JSONField, ListCharField,
+    ListTextField, NullBit1BooleanField, SetCharField, SetTextField,
+    SizedBinaryField, SizedTextField,
 )