+++ django_mysql/models/fields/__init__.py
@@ -3,6 +3,7 @@ from django_mysql.models.fields.bit import (  # noqa
 )
 from django_mysql.models.fields.dynamic import DynamicField  # noqa
 from django_mysql.models.fields.enum import EnumField  # noqa
+from django_mysql.models.fields.json import JSONField  # noqa
 from django_mysql.models.fields.lists import (  # noqa
     ListCharField, ListTextField
 )